Set Filter for all auctions
